+++ client/src/app/shared/shared.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { HeaderComponent, FooterComponent } from '@app/shared';
+import { HeaderComponent, FooterComponent, HeaderShadowDirective } from '@app/shared';
 
 @NgModule({
   imports: [
     CommonModule
   ],
-  declarations: [HeaderComponent, FooterComponent],
-  exports: [HeaderComponent, FooterComponent]
+  declarations: [HeaderComponent, FooterComponent, HeaderShadowDirective],
+  exports: [HeaderComponent, FooterComponent, HeaderShadowDirective]
 })
 export class SharedModule { }